requirements.txt
ansible
docopt
pycrypto
dopy
